# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "SolidCAD Variant Customer API documentation",
    "version": "1.0.0",
    "title": "SolidCAD Variant Customer API",
    "termsOfService": "https://www.solidcad.ca/privacy-policy/",
    "contact": {
      "email": "apiteam@solidcad.ca"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:1002",
  "basePath": "/api/v1/",
  "tags": [
    {
      "name": "Variant",
      "description": "Everything about Variant APIs",
      "externalDocs": {
        "description": "Find out more",
        "url": "https://www.solidcad.ca/products/solidcad-products/variant/"
      }    
    }
  ],
  "schemes": [
     "http"
  ],
  "paths": {
    "/contactUs": {
      "post": {
        "tags": [
          "Variant"
        ],
        "summary": "Create a new Enquiry and Send Email Confirmation",
        "description": "",
        "operationId": "contactUs",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Fill the all the required fields and received email",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Variant"
            }
          }
        ],
        "responses": {
          "201":{
            "schema":{
              "type":"object",
              "properties":{
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "message":{
                  "type":"string",
                  "example": "Thanks you for reaching SolidCad Team!!We will get back to you asap!!"
                },
                "response":{
                  "type":"object",
                  "example":"Document"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Tag": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
    "Variant": {
      "type": "object",
      "required": [
        "firstName",
        "lastName",
        "email",
        "message"
      ],
      "properties": {
        "firstName": {
          "type":"string",
          "example":"Alex"
        },
        "lastName": {
          "type": "string",
          "example": "doggie"
        },
        "email":{
            "type":"email",
            "example":"alex.doggie@gmail.com"
        },
        "message": {
            "type": "string",
            "example": "Hello I want to know more about solidCAD Variant Product!"
        }
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}